feat: write readme.md, add error handling cases if date is alphabets, remove errors by testing with different cases
